+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.4.0] - 2025-11-16
+
 ### Added
 
 - Custom Docker image backend support as an alternative to Restic
@@ -99,7 +101,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Docker event listener for dynamic discovery
 - Configuration via config.yml and Docker labels
 
-[Unreleased]: https://github.com/polarfoxDev/marina/compare/v0.3.1...HEAD
+[Unreleased]: https://github.com/polarfoxDev/marina/compare/v0.4.0...HEAD
+[0.4.0]: https://github.com/polarfoxDev/marina/compare/v0.3.1...v0.4.0
 [0.3.1]: https://github.com/polarfoxDev/marina/compare/v0.3.0...v0.3.1
 [0.3.0]: https://github.com/polarfoxDev/marina/compare/v0.2.1...v0.3.0
 [0.2.1]: https://github.com/polarfoxDev/marina/compare/v0.2.0...v0.2.1